move visualization CLI into bin
