+++ fastestimator/util/util.py
@@ -1,5 +1,11 @@
-import tensorflow as tf
+import argparse
+import os
+import re
+import string
 import subprocess
+
+import tensorflow as tf
+
 
 def get_num_GPU():
     """
@@ -40,3 +46,85 @@ def convert_tf_dtype(datatype):
                     "float32": tf.float32,
                     "float64": tf.float64}
     return datatype_map[datatype]
+
+
+def strip_suffix(target, suffix):
+    """
+    Remove the given suffix from the target if it is present there
+
+    Args:
+        target: A string to be formatted
+        suffix: A string to be removed from 'target'
+
+    Returns:
+        The formatted version of 'target'
+    """
+    if suffix is None or target is None:
+        return target
+    s_len = len(suffix)
+    if target[-s_len:] == suffix:
+        return target[:-s_len]
+    return target
+
+
+def prettify_metric_name(metric):
+    """
+    Add spaces to camel case words, then swap _ for space, and capitalize each word
+
+    Args:
+        metric: A string to be formatted
+
+    Returns:
+        The formatted version of 'metric'
+    """
+    return string.capwords(re.sub("([a-z])([A-Z])", r"\g<1> \g<2>", metric).replace("_", " "))
+
+
+def remove_blacklist_keys(dic, blacklist):
+    """
+    A function which removes the blacklisted elements from a dictionary
+
+    Args:
+        dic: The dictionary to inspect
+        blacklist: keys to be removed from dic if they are present
+
+    Returns:
+        None
+
+    Side Effects:
+        Entries in dic may be removed
+    """
+    if blacklist is None:
+        return
+    for elem in blacklist:
+        dic.pop(elem, None)
+
+
+def is_number(s):
+    """
+    Args:
+        s: A string
+    Returns:
+        True iff the string represents a number
+    """
+    try:
+        float(s)
+        return True
+    except ValueError:
+        return False
+
+
+class SaveAction(argparse.Action):
+    """
+    A custom save action which is used to populate a secondary variable inside of an exclusive group. Used if this file
+        is invoked directly during argument parsing.
+    """
+
+    def __init__(self, option_strings, dest, nargs='?', **kwargs):
+        if '?' != nargs:
+            raise ValueError("nargs must be \'?\'")
+        super(SaveAction, self).__init__(option_strings, dest, nargs, **kwargs)
+
+    def __call__(self, parser, namespace, values, option_string=None):
+        setattr(namespace, self.dest, True)
+        setattr(namespace, self.dest + '_dir', values if values is None else os.path.join(values, ''))
